// Ticket Tout/migrations/snapshots/4edcae4c337f670648995f1167f6043364cea1c6973e9465502baf5f37ae4ce0/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Administration: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly CategoryEntreprise: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly category: CodecTypes['pg/text@1']['output'];
    };
    readonly Document: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly storageKey: CodecTypes['pg/text@1']['output'];
      readonly mimeType: CodecTypes['pg/text@1']['output'];
      readonly size: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Entreprise: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly siret: CodecTypes['pg/text@1']['output'];
      readonly kbisId: CodecTypes['pg/text@1']['output'];
      readonly addresse: CodecTypes['pg/text@1']['output'];
      readonly postalCode: CodecTypes['pg/text@1']['output'];
      readonly agentId: CodecTypes['pg/int4@1']['output'];
      readonly motifId: CodecTypes['pg/int4@1']['output'];
      readonly verified: CodecTypes['pg/bool@1']['output'];
      readonly categoryId: CodecTypes['pg/int4@1']['output'];
      readonly isPartenaire: CodecTypes['pg/bool@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly HistoriqueTransactionEntreprise: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly transactionId: CodecTypes['pg/text@1']['output'];
      readonly entrepriseId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly HistoriqueTransactionSalarie: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly transactionId: CodecTypes['pg/text@1']['output'];
      readonly salarieId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly MotifValidationEntreprise: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly motif: CodecTypes['pg/text@1']['output'];
    };
    readonly Transaction: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly receveurId: CodecTypes['pg/text@1']['output'];
      readonly envoyeurId: CodecTypes['pg/text@1']['output'];
      readonly montant: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly UserRole: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly role: CodecTypes['pg/text@1']['output'];
    };
    readonly Users: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly surname: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly roleId: CodecTypes['pg/int4@1']['output'];
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly expiredAt: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly documentId: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Administration: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly CategoryEntreprise: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly category: CodecTypes['pg/text@1']['input'];
    };
    readonly Document: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly storageKey: CodecTypes['pg/text@1']['input'];
      readonly mimeType: CodecTypes['pg/text@1']['input'];
      readonly size: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Entreprise: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly siret: CodecTypes['pg/text@1']['input'];
      readonly kbisId: CodecTypes['pg/text@1']['input'];
      readonly addresse: CodecTypes['pg/text@1']['input'];
      readonly postalCode: CodecTypes['pg/text@1']['input'];
      readonly agentId: CodecTypes['pg/int4@1']['input'];
      readonly motifId: CodecTypes['pg/int4@1']['input'];
      readonly verified: CodecTypes['pg/bool@1']['input'];
      readonly categoryId: CodecTypes['pg/int4@1']['input'];
      readonly isPartenaire: CodecTypes['pg/bool@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly HistoriqueTransactionEntreprise: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly transactionId: CodecTypes['pg/text@1']['input'];
      readonly entrepriseId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly HistoriqueTransactionSalarie: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly transactionId: CodecTypes['pg/text@1']['input'];
      readonly salarieId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly MotifValidationEntreprise: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly motif: CodecTypes['pg/text@1']['input'];
    };
    readonly Transaction: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly receveurId: CodecTypes['pg/text@1']['input'];
      readonly envoyeurId: CodecTypes['pg/text@1']['input'];
      readonly montant: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly UserRole: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly role: CodecTypes['pg/text@1']['input'];
    };
    readonly Users: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly surname: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly roleId: CodecTypes['pg/int4@1']['input'];
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly expiredAt: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly documentId: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly administration: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly categoryEntreprise: {
      readonly category: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
    };
    readonly document: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly mimeType: CodecTypes['pg/text@1']['output'];
      readonly size: CodecTypes['pg/int4@1']['output'];
      readonly storageKey: CodecTypes['pg/text@1']['output'];
    };
    readonly entreprise: {
      readonly addresse: CodecTypes['pg/text@1']['output'];
      readonly agentId: CodecTypes['pg/int4@1']['output'];
      readonly categoryId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly isPartenaire: CodecTypes['pg/bool@1']['output'];
      readonly kbisId: CodecTypes['pg/text@1']['output'];
      readonly motifId: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly postalCode: CodecTypes['pg/text@1']['output'];
      readonly siret: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly verified: CodecTypes['pg/bool@1']['output'];
    };
    readonly historiqueTransactionEntreprise: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly entrepriseId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly transactionId: CodecTypes['pg/text@1']['output'];
    };
    readonly historiqueTransactionSalarie: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly salarieId: CodecTypes['pg/text@1']['output'];
      readonly transactionId: CodecTypes['pg/text@1']['output'];
    };
    readonly motifValidationEntreprise: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly motif: CodecTypes['pg/text@1']['output'];
    };
    readonly transaction: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly envoyeurId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly montant: CodecTypes['pg/int4@1']['output'];
      readonly receveurId: CodecTypes['pg/text@1']['output'];
    };
    readonly userRole: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly role: CodecTypes['pg/text@1']['output'];
    };
    readonly users: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly documentId: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly expiredAt: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly roleId: CodecTypes['pg/int4@1']['output'];
      readonly surname: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;